Refactor Kafka config
